tests: removed pending_mediated_transfer
The pending_mediated_transfer utility was going around the RaidenService
api and injecting state changes directly in the write-ahead-log. This
approach does produce correct results in the state logic, but not in the
RaidenService, since the AsyncResults are not properly set.

This change removed the pending_mediated_transfer in favor of a special
event handler, which allows the tests to decide which secret requests
are sent. Having the same end effect of pending_mediated_transfer while
properly setting up the RaidenService state.