Makes HTTPExecutor wait for no process to be listening on port on stop
Sometimes, between kill/stop and the synapse socket which listens in the port
to be closed, there may be a small delay, which prevented the next
synapse instance to run. This commit includes the check for a running
process in the given port for running(), so stop() will wait for no
process to be listening on that port, and then kill-9 it if it lasts
more than 10s. In practice, it should just wait a little more before
exiting. Synapse test server is used only in tests, with :memory: db, so
no need to be kind to it