delayed expired lock removal
the receiver of a lock expired message must wait for the confirmation
blocks before accepting the message, otherwise the secret could have
been registered on a reorg which is unknown to the node.

The sender must take this into account, and only send the remove expired
lock after the confirmation blocks has elapsed, with a few caveats:

- the confirmation blocks is a local configuration, meaning the sender
may have a lower value than the receiver, and the queue will be stalled
until the target's node requirement is met.
- waiting for confirmation blocks is not sufficient, because the
sender's blockchain may be "more in synch" then the target's chain. IOW,
the target's chain may be slightly behind from the chain's HEAD, and
from its perspective the lock has not yet expired.

Because of the second point above, the sender (as an initiator and/or
mediator), is now waiting for confirmation blocks * 2.