proxy: added a check for recoverable condition
update transfer has a limited time frame in which it can be called,
which is equal to the settlement window. Because this window is time
dependent it is always possible that a transaction is not mined in time.
This is the worst case but it is possible. To prevent the node from
crashing because of a known race condition the exception was set as
recoverable error.