test_payment_channel_proxy_basics works.
Using test_payment_channel_proxy_basics() as a test made a lot of changes to
payment channel and token network proxy.

Still a lot of work to do. Main conflict point at the moment is that we have
functions that send transactions to the blockchain just like the
`new_netting_channel()` does. Before this commit it was not working but in this
commit I changed the block hash right after the open transaction was mined in
the same function.

The same would need to be done in all of the other places in the proxies where
inside the same function a transaction is sent and mined, because this polls and
waits for CONFIRMATION blocks and thus we need to change the block hash by which
we query the chain for the subsequent function calls.