Fix events_for_onchain_secretreveal_if_dangerzone
The function events_for_onchain_secretreveal_if_dangerzone assume that
transfer_pairs always had entries, which is not a valid assumption after
the changes for the remove expired lock. Not the mediate task must be
kept alive until the lock is properly removed from the channel state,
and it's possible for the mediator to fail to forward any of the
received transfers, which makes the list empty.