Fix restore_to_state_change logging
The function `get_snapshot_closest_to_state_change` always returns a
tuple, which may be `(None, 0)`. In the previous code this return value
would be check for truthiness, which even though the values in the tuple
evaluate to False, the tuple itself evaluate to True because it has
elements. The end result was just improper logging.

This also improves the logging and adds assertions to easy debugging.