Forbid invalid upgrades
When a database version upgrade happens, the user must take action and
exit all the token networks and rename the existing database file. This
should not happen often, as the database format is choosen to be
upgradable, but to avoid problems with testing realeases and possible
future breaking changes a check was added.