Fix Refund transfer serialization
The shape of messages which balance proofs changes depending if it
contains a lock or not. Messages like Unlock, LockExpired and
DirectTransfer have a balance proof directly in the object itself, while
messages like LockedTransfer and RefundTransfer have their balance proof
inside a transfer object, which also contains the lock.

This difference in the shape of the objects would also require different
queries to the database. In order to not have such queries, the
serialization of the objects duplicates information, and copies the
balance proofs from the inner transfer object to the top level, allowing
a single query to work with all the above messages and their state
changes counter part.

However, such strategy is error prone, and we indeed forgot to properly
serialize two state changes, namely ReceiveTransferRefundCancelRoute and
ReceiveTransferRefund which are used for refund transfers. In this case,
if a refund transfer was the last transfer received for a given channel,
and this channel was closed, the call to settlement fails, because the
corresponding balance proof cannot be found.

This commit adds a new marker for the state changes which have a balance
proof, fixes the serialization of the refund state changes, and adds a
unit test for the current state changes which have a balance proof.