Test: Factories with most arguments optional
Note however that this gives no guarantee on the correctness or
availability of the data, meaning it's perfectly fine for the factory to
fail because the constructed objet would be inconsistent, or for the
test to fail if the values are inconsistent.

This is a big change in respect to the previous approach, which tried to
do exactly the opposite, and force the test writer to choose appropriate
values, however this is a step torwards making it possible to use value
generation (because of the type annotations and name consolidation).

This also introduces a new object named EMPTY, since `None` may be a
valid value in some cases.