Add make mypy check in CircleCI builds
Although mypy still returns many error messages, for some files, some
kinds of errors have been eliminated.  I'm adding a partial check in
CircleCI builds.  The plan is to make this check more and more powerful
until mypy reports no errors.

This is a part of https://github.com/raiden-network/raiden/issues/3040.

But Travis needs not to run on this commit so seriously
[no ci integration]