Re-use previous matrix login credentials
Server and account are matched, and if different, it's skipped.
A '/sync' call is done to test validity of the credentials, and if it
fails with 401: 'Unrecognized access token', a new login is done.
As we're using previous tokens, don't logout on stop, fixing #3130
Implements a long-standing TODO.
This new logic is tested by tests/integration/test_send_queued_messages.py