After refund transfer with no new routes don't delete payment task
After the initiator receives a RefundTransfer and has no new routes to
forward the payment the payment should fail but the payment task
should not be removed.

The reasoning behind this is that the mediator who sent us the refund
transfer does not know that the payment is canceled and keeps it on
his state until lock expiration. So he will eventually also send us a
`LockExpired` message when the lock expires.

We need to have the initiator payment task alive at that point or else
we will not process the LockExpired and see the nonce mismatch
problems described in #3146 which will lead to the channel getting stuck.

Related to the comment in
https://github.com/raiden-network/raiden/issues/3146#issuecomment-447378046

Fix #3146