Handle the Parity Transaction already imported error
Fix #3211 for parity.

What we do is:

1. When we get the transaction already imported error we get all the
pending transactions from the pool via `parity_allTransactions`.
2. Search these transactions for one originating from our node and
with the expected nonce.
3. i. If the transaction is found we get the hash and proceed as usual
with polling it.
   ii. If it's not found we handle this as the geth
   `TransactionAlreadyPending` error and crash raiden.

3-ii can only happen if the transaction gets out of the pool between
the time we get the already imported error and the time we query it. The only reason that 3-ii does not handle the situation gracefully is
due to the fact that there is no way to get the transaction hash and
at that point in the client we need a transaction hash so we are left
with no choice.