Poll block events once more after first alarm task run
Fixes #3216

In some cases there can be a pending transaction left in the
transaction queues after the first run of the alarm task that would
have been invalidated if we get the events generated by all the
transactions sent during the first run of the alarm task.

This commit addresses this by adding a poll of the new block events
between the first run of the alarm task and the initialization of the
transaction queues.