simplified lock acquision in the python api
- removed locks from the payment_channel proxy, since the token network
already has locks in place.
- Channel operations are no longer raising a ChannelBusy exception if
there is another concurrent operation. The new transaction will just
wait for the pending one to finish, and once the lock is held the
preconditions checks if the transaction is still valid.
    - This means that if a user request a channel close and there is a
    deposit in flight, the close will not raise an error but wait for
    the deposit to finish and then the close is run.
    - The same happens for the deposit waiting for a close, but the
    deposit is not sent once the lock is acquired because of the
    preconditions checks.

The above means we no longer need to acquire the locks for all channels
before doing a batch close, we can just dispatch the action close and
wait for the locks to be freed, and eventually the channels will close.