Optimization: Removed redundant deepcopy
The previous code did a deepcopy before every state change dispatch.
This was done to prevent concurrent changes to the node's state, which
can lead to nasty bugs. That, however, is a very expensive and wasteful
operations. This commit changes the access pattern of the node state,
instead of having direct access to the `NodeState` instance, now the
code is expecte to go through the `get_current_state` function from the
`WriteAheadLog` object, which will then do a copy. This has the same
effect as before, were a consumer can freely look into the node state
and be sure it will not be changed by another thread, while doing just
the amount of work that is necessary.

This also changes the type signature of the `StateManager` to require an
initial state, this removes lots of checks for `None` from the code,
since now it is know there is always a state set in the manager.